requirements.txt
Flask
flask-cors
tensorflow-cpu
tf-keras
opencv-python
pillow
numpy
werkzeug
gunicorn
